encourage use of version_info() in issues
