+++ .github/ISSUE_TEMPLATE/bug.md
@@ -6,15 +6,16 @@ labels: bug
 
 # Bug
 
-Please complete:
-* OS: **?**
-* Python version `import sys; print(sys.version)`: **?**
-* Pydantic version `import pydantic; print(pydantic.VERSION)`: **?**
+Output of `python -c "import pydantic.utils; print(pydantic.utils.version_info())"`:
+```
+...
+```
+<!-- or if you're using pydantic before v1.3, manually add os, python version and pydantic version -->
 
-**Please read the [docs](https://pydantic-docs.helpmanual.io/) and search through issues to
-confirm your bug hasn't already been reported.**
+<!-- Please read the [docs](https://pydantic-docs.helpmanual.io/) and search through issues to
+confirm your bug hasn't already been reported. -->
 
-Where possible please include a self contained code snippet describing your bug:
+<!-- Where possible please include a self contained code snippet describing your bug: -->
 
 ```py
 import pydantic
