+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -6,15 +6,16 @@ labels: feature request
 
 # Feature Request
 
-Please complete:
-* OS: **?**
-* Python version `import sys; print(sys.version)`: **?**
-* Pydantic version `import pydantic; print(pydantic.VERSION)`: **?**
+Output of `python -c "import pydantic.utils; print(pydantic.utils.version_info())"`:
+```
+...
+```
+<!-- or if you're using pydantic before v1.3, manually add os, python version and pydantic version -->
 
-**Please read the [docs](https://pydantic-docs.helpmanual.io/) and search through issues to
-confirm your feature hasn't been asked for before, or already implemented.**
+<!-- Please read the [docs](https://pydantic-docs.helpmanual.io/) and search through issues to
+confirm your feature hasn't been asked for before, or already implemented. -->
 
-Where possible please include a self contained code snippet describing your feature request:
+<!-- Where possible please include a self contained code snippet describing your feature request: -->
 
 ```py
 import pydantic
