+++ .github/ISSUE_TEMPLATE/question.md
@@ -6,15 +6,16 @@ labels: question
 
 # Question
 
-Please complete:
-* OS: **?**
-* Python version `import sys; print(sys.version)`: **?**
-* Pydantic version `import pydantic; print(pydantic.VERSION)`: **?**
+Output of `python -c "import pydantic.utils; print(pydantic.utils.version_info())"`:
+```
+...
+```
+<!-- or if you're using pydantic before v1.3, manually add os, python version and pydantic version -->
 
-**Please read the [docs](https://pydantic-docs.helpmanual.io/) and search through issues to
-confirm your question hasn't already been answered.**
+<!-- Please read the [docs](https://pydantic-docs.helpmanual.io/) and search through issues to
+confirm your question hasn't already been answered. -->
 
-Where possible please include a self contained code snippet describing your question:
+<!-- Where possible please include a self contained code snippet describing your question: -->
 
 ```py
 import pydantic
