+++ docs/contributing.md
@@ -4,13 +4,15 @@ We'd love you to contribute to *pydantic*!
 
 Questions, feature requests and bug reports are all welcome as issues.
 
-To make it as simple as possible for us to help you please include the output of the following call in your issue:
+To make it as simple as possible for us to help you, please include the output of the following call in your issue:
 
 ```bash
 python -c "import pydantic.utils; print(pydantic.utils.version_info())"
 ```
+If you're using *pydantic* prior to **v1.3** (when `version_info()` was added), please manually include OS, python
+version and pydantic version.
 
-Please try to always include the above output unless you're unable to install *pydantic* or **know** it's not relevant
+Please try to always include the above unless you're unable to install *pydantic* or **know** it's not relevant
 to your question or feature request.
 
 ## Pull Requests
