+++ pydantic/utils.py
@@ -19,13 +19,29 @@ from typing import (
 )
 
 from .typing import AnyType, display_as_type
+from .version import version_info
 
 if TYPE_CHECKING:
     from .main import BaseModel  # noqa: F401
     from .typing import AbstractSetIntStr, DictIntStrAny, IntStr, ReprArgs  # noqa: F401
     from .dataclasses import DataclassType  # noqa: F401
 
-KeyType = TypeVar('KeyType')
+__all__ = (
+    'import_string',
+    'sequence_like',
+    'validate_field_name',
+    'lenient_issubclass',
+    'in_ipython',
+    'deep_update',
+    'update_not_none',
+    'almost_equal_floats',
+    'get_model',
+    'PyObjectStr',
+    'Representation',
+    'GetterDict',
+    'ValueItems',
+    'version_info',  # required here to match behaviour in v1.3
+)
 
 
 def import_string(dotted_path: str) -> Any:
@@ -64,9 +80,6 @@ def truncate(v: Union[str], *, max_len: int = 80) -> str:
     return v
 
 
-ExcType = Type[Exception]
-
-
 def sequence_like(v: AnyType) -> bool:
     return isinstance(v, (list, tuple, set, frozenset, GeneratorType))
 
@@ -97,6 +110,9 @@ def in_ipython() -> bool:
         return False
     else:  # pragma: no cover
         return True
+
+
+KeyType = TypeVar('KeyType')
 
 
 def deep_update(mapping: Dict[KeyType, Any], updating_mapping: Dict[KeyType, Any]) -> Dict[KeyType, Any]:
@@ -332,31 +348,3 @@ class ValueItems(Representation):
 
     def __repr_args__(self) -> 'ReprArgs':
         return [(None, self._items)]
-
-
-def version_info() -> str:
-    import platform
-    import sys
-    from importlib import import_module
-    from pathlib import Path
-
-    from .main import compiled
-    from .version import VERSION
-
-    optional_deps = []
-    for p in ('typing-extensions', 'email-validator', 'devtools'):
-        try:
-            import_module(p.replace('-', '_'))
-        except ImportError:
-            continue
-        optional_deps.append(p)
-
-    info = {
-        'pydantic version': VERSION,
-        'pydantic compiled': compiled,
-        'install path': Path(__file__).resolve().parent,
-        'python version': sys.version,
-        'platform': platform.platform(),
-        'optional deps. installed': optional_deps,
-    }
-    return '\n'.join('{:>30} {}'.format(k + ':', str(v).replace('\n', ' ')) for k, v in info.items())
